Move variant_stock.rb to concerns/
