+++ app/models/spree/variant_decorator.rb
@@ -1,6 +1,6 @@
 require 'open_food_network/enterprise_fee_calculator'
 require 'open_food_network/variant_and_line_item_naming'
-require 'open_food_network/variant_stock'
+require 'concerns/variant_stock'
 require 'open_food_network/products_cache'
 
 Spree::Variant.class_eval do
@@ -10,7 +10,7 @@ Spree::Variant.class_eval do
   # removing the Spree method to prevent error.
   remove_method :options_text if instance_methods(false).include? :options_text
   include OpenFoodNetwork::VariantAndLineItemNaming
-  include OpenFoodNetwork::VariantStock
+  include VariantStock
 
   has_many :exchange_variants
   has_many :exchanges, through: :exchange_variants
