+++ spec/models/concerns/variant_stock_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe OpenFoodNetwork::VariantStock do
+describe VariantStock do
   let(:variant) { create(:variant) }
 
   describe '#after_save' do
